+++ include/sensors.h
@@ -13,7 +13,7 @@ public:
 
     Gyro(I2C *i);
     int init(Range r); // 0 on success
-    int read(double *v); // 0 on success
+    int read(float *v); // 0 on success
 
     static const uint8_t address = 0xD6;
     static const uint8_t register1 = 0x20;
@@ -37,7 +37,7 @@ public:
 
     Acc(I2C *i);
     int init(Range r); // 0 on success
-    int read(double *v); // 0 on success
+    int read(float *v); // 0 on success
 
     static const uint8_t address = 0x32;
     static const uint8_t register1 = 0x20;
